+++ test/io_tests/connect_test.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 
 import pytest
 from xknx import XKNX
-from xknx.io import Connect, UDPClient
+from xknx.io import UDPClient
+from xknx.io.request_response import Connect
 from xknx.knxip import (
     HPAI,
     ConnectRequest,